Add in test-case generation engine
We can now feed in yaml files that will generate many variations
of a given test case.

These tests are designed to be written in a more-generic fashion and
to contain the logic to test several variations based on input variables.
Thus one test module and a yaml file can generate many individual test cases
to be run.